+++ apps/test/unit/lib/kits/maker/ui/DiscountCodeSchoolChoiceTest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {assert} from '../../../util/configuredChai';
-import DiscountCodeSchoolChoice from '@cdo/apps/templates/maker/DiscountCodeSchoolChoice';
+import {assert} from '../../../../../util/configuredChai';
+import DiscountCodeSchoolChoice from '@cdo/apps/lib/kits/maker/ui/DiscountCodeSchoolChoice';
 
 describe('DiscountCodeSchoolChoice', () => {
   const defaultProps = {
