+++ apps/test/unit/lib/kits/maker/ui/EligibilityChecklistTest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {assert} from '../../../util/configuredChai';
-import EligibilityChecklist from '@cdo/apps/templates/maker/EligibilityChecklist';
+import {assert} from '../../../../../util/configuredChai';
+import EligibilityChecklist from '@cdo/apps/lib/kits/maker/ui/EligibilityChecklist';
 import {Status} from '@cdo/apps/lib/ui/ValidationStep';
 
 describe('EligibilityChecklist', () => {
